move logging.cfg to logging folder
