+++ great_expectations_cloud/agent/cli.py
@@ -5,7 +5,7 @@ import dataclasses as dc
 import logging
 import pathlib
 
-from great_expectations_cloud.logging_cfg import LogLevel, configure_logger
+from great_expectations_cloud.logging.logging_cfg import LogLevel, configure_logger
 
 LOGGER = logging.getLogger(__name__)
